# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ScrollUniversity API
  description: |
    Complete API documentation for the ScrollUniversity platform - a revolutionary Christian educational platform 
    that combines divine revelation with cutting-edge technology.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 5 requests per 15 minutes
    - API endpoints: 1000 requests per hour
    
    ## Response Format
    All responses follow this structure:
    ```json
    {
      "success": true,
      "data": { ... },
      "error": null
    }
    ```
  version: 1.0.0
  contact:
    name: ScrollUniversity API Support
    email: api@scrolluniversity.org
  license:
    name: Proprietary
    url: https://scrolluniversity.org/license
servers:
  - url: https://api.scrolluniversity.org/v1
    description: Production server
  - url: https://staging-api.scrolluniversity.org/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [STUDENT, FACULTY, ADMIN]
        avatar:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        faculty:
          type: string
        instructor:
          $ref: '#/components/schemas/User'
        ScrollGoldCost:
          type: number
        usdPrice:
          type: number
        duration:
          type: number
          description: Duration in hours
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        modules:
          type: array
          items:
            $ref: '#/components/schemas/CourseModule'
    CourseModule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        order:
          type: number
        lectures:
          type: array
          items:
            $ref: '#/components/schemas/Lecture'
    Lecture:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [video, live, interactive]
        videoUrl:
          type: string
          format: uri
        duration:
          type: number
        transcript:
          type: string
...
paths:
...
  /courses:
    get:
      tags:
        - Courses
      summary: List all courses
      description: Get paginated list of available courses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: faculty
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      courses:
                        type: array
                        items:
                          $ref: '#/components/schemas/Course'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
  /courses/{courseId}:
    get:
      tags:
        - Courses
      summary: Get course details
      description: Retrieve detailed information about a specific course
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Course'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []